requirements.txt
editdistance==0.6.2
elasticsearch==8.9.0
faiss-cpu==1.7.4
farm-haystack==1.20.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
jsonlines==4.0.0
mlflow==2.7.1
ngram==4.0.3
nltk==3.6.3
num2words==0.5.12
numpy==1.24.4
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.4.0
torchvision==0.15.2
transformers==4.32.1